backend/pyproject.toml
[tool.poetry]
name = "roskos-task"
version = "0.1.0"
description = ""
authors = ["RussianProgram <kurban.kerimhanov@gmail.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.2"
djangorestframework = "^3.14.0"
django-filter = "^22.1"
psycopg2-binary = "^2.9.5"
django-cors-headers = "^3.13.0"
black = "^22.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
